add `Solana` coin
1. add `SOL` transfer
2. add sol token transfer
3. add associated token create
4. add soft impl for solana
5. add test code

+++ src/context/BaseContext.cpp
@@ -116,7 +116,7 @@ JUB_RV BaseContext::SetTimeout(const JUB_UINT16 timeout) {
 std::string BaseContext::_FullBip44Path(const BIP44_Path& path) {
 
     return _mainPath + "/"
-        + std::to_string(path.change) + "/"
+        + std::to_string((uint32_t)path.change) + "/"
         + std::to_string(path.addressIndex);
 }
